[#750] Rename: interface and default implementation of health check server.
Renamed interface HealthCheckRegistration to HealthCheckServer and class
HealthCheckServer to VertxBasedHealthCheckServer.
