+++ service-base/src/main/java/org/eclipse/hono/service/AbstractServiceBase.java
@@ -40,7 +40,7 @@ import io.vertx.ext.healthchecks.HealthCheckHandler;
  */
 public abstract class AbstractServiceBase<T extends ServiceConfigProperties> extends ConfigurationSupportingVerticle<T> implements HealthCheckProvider {
 
-    private HealthCheckRegistration healthCheckServer = new NoopHealthCheckServer();
+    private HealthCheckServer healthCheckServer = new NoopHealthCheckServer();
 
     /**
      * A logger to be shared with subclasses.
@@ -68,13 +68,13 @@ public abstract class AbstractServiceBase<T extends ServiceConfigProperties> ext
     }
 
     /**
-     * Sets the HealthCheckRegistration to be used.
-     * 
-     * @param healthCheckServer The HealthCheckRegistration.
+     * Sets the health check server for this application.
+     *
+     * @param healthCheckServer The health check server.
      * @throws NullPointerException if healthCheckServer is {@code null}.
      */
     @Autowired(required = false)
-    public void setHealthCheckServer(final HealthCheckRegistration healthCheckServer) {
+    public void setHealthCheckServer(final HealthCheckServer healthCheckServer) {
         this.healthCheckServer = Objects.requireNonNull(healthCheckServer);
     }
